+++ NEWS.md
@@ -25,7 +25,7 @@ Development
 - Scroll fixes ([#14704](https://github.com/CartoDB/cartodb/issues/14704), [#14703](https://github.com/CartoDB/cartodb/issues/14703))
 - Fix bug with Shift and Click ([CartoDB/product#279](https://github.com/CartoDB/product/issues/279))
 - Improve shift and click behavior ([CartoDB/product#278](https://github.com/CartoDB/product/issues/278))
-- Add new Ghost tables task to be used byt the TIS([#14731](https://github.com/CartoDB/cartodb/pull/14731))
+- Add new Ghost tables resque task with username as parameter([#14731](https://github.com/CartoDB/cartodb/pull/14731))
 - Add developer center documentation folder to doc [14710](https://github.com/CartoDB/cartodb/pull/14710)